// src/services/types/data.ts
  export interface IRegisterUserRequest {
    readonly name: string;
    readonly email: string;
    readonly password: string;
  }

  export interface ILoginUserRequest {
    readonly email: string;
    readonly password: string;
  }
// ...
  export interface IResetPasswordRequest {
    readonly password: string;
    readonly token: string;
  }
// ...
  export interface ISetLoginFormValue {
    readonly type: typeof USER_LOGIN_FORM_SET_VALUE;
    readonly field: string;
    readonly value: string;
  }
  
  export interface IUserRegisterFormSubmitFailed {
    readonly type: typeof USER_REGISTER_FORM_SUBMIT_FAILED;
  }
  
  export interface IUserRegisterFormSubmitSuccess {
    readonly type: typeof USER_REGISTER_FORM_SUBMIT_SUCCESS;
  }

  export interface IUserRegisterFormSubmit {
    readonly type: typeof USER_REGISTER_FORM_SUBMIT;
  }


  export interface ISetRegisterFormValue {
    readonly type: typeof USER_REGISTER_FORM_SET_VALUE;
    readonly field: string;
    readonly value: string;
  }

  export interface IUserAccessDenied {
    readonly type: typeof USER_ACCESS_DENIED;
  }

  export interface IUserAccessAllowed {
    readonly type: typeof USER_ACCESS_ALLOWED;
    readonly payload: IUser;
  }

  export interface ISetForgotPasswordFormValue {
    readonly type: typeof FORGOT_PASSWORD_FORM_SET_VALUE;
    readonly field: string;
    readonly value: string;
  }

  export interface IForgotPasswordFormSubmit {
    readonly type: typeof FORGOT_PASSWORD_FORM_SUBMIT;
  }
  
  export interface IForgotPasswordFormSubmitSuccess {
    readonly type: typeof FORGOT_PASSWORD_FORM_SUBMIT_SUCCESS;
  }
  
  export interface IForgotPasswordFormSubmitFailed {
    readonly type: typeof FORGOT_PASSWORD_FORM_SUBMIT_FAILED;
  }

  export interface ISetResetPasswordFormValue {
    readonly type: typeof RESET_PASSWORD_FORM_SET_VALUE;
    readonly field: string;
    readonly value: string;
  }

  export interface IResetPasswordFormSubmit {
    readonly type: typeof RESET_PASSWORD_FORM_SUBMIT;
  }
// ...
  export interface IChangeUserDataRequest {
    readonly email: string;
    readonly name: string;
  }